+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-scripts": "1.0.7"
+    "react-scripts": "^1.1.4"
   },
   "dependencies": {
     "jquery": "^3.2.1",
