+++ src/actions/index.js
@@ -12,41 +12,3 @@ export const staffFetchDataSuccess = (staff) => ({
   type: 'STAFF_FETCH_DATA_SUCCESS',
   staff
 })
-
-export const fetchStaff = (url) => {
-  return (dispatch) =>  {
-    dispatch(isLoading(true))
-    fetch(url)
-      .then(response => {
-        if (!response.ok) {
-          throw Error(response.statusText)
-        }
-        dispatch(isLoading(false))
-        return response
-    })
-    .then(response => response.json())
-    .then(data => dispatch(fetchBios(data.bio)))
-    .then(staff => dispatch(staffFetchDataSuccess(staff)))
-    .catch(() => dispatch(hasErrored(true)))
-  }
-}
-
-export const fetchBios = (staffArray) => {
-  return (dispatch) => {
-    dispatch(isLoading(true))
-    const unresolvedPromises = staffArray.map(staffMember => {
-      return fetch(staffMember.info)
-      .then(response => {
-        if (!response.ok) {
-          throw Error(response.statusText)
-        }
-        dispatch(isLoading(false))
-        return response
-      })
-      .then(response => response.json())
-      .then(data => ({...data, name: staffMember.name}))
-      .catch(() => dispatch(hasErrored(true)))
-    })
-    return Promise.all(unresolvedPromises);
-  }
-}
